+++ FusionIIIT/applications/hr2/api/serializers.py
@@ -183,12 +183,3 @@ class LeaveBalanace_serializer(serializers.ModelSerializer):
 
     def create(self, validated_data):
         return LeaveBalance.objects.create(**validated_data)
-
-
-# class Deignations(serializers.ModelSerializer):
-#     class Meta:
-#         model = Deignations
-#         fields = '__all__'
-
-#     def create(self,validated_data):
-#         return
