--- FusionIIIT/applications/hr2/api/views/form_views.py
@@ -1,15 +1,0 @@
-from .ltc_views import LTC
-from .cpda_views import CPDAAdvance, CPDAReimbursement
-from .leave_views import Leave
-from .appraisal_views import Appraisal
-from .management_views import (
-    FormManagement,
-    GetFormHistory,
-    TrackProgress,
-    FormFetch,
-    CheckLeaveBalance,
-    DropDown,
-    UserById,
-    ViewArchived,
-    GetOutbox,
-)
