--- FusionIIIT/applications/hr2/normal.py
@@ -1,36 +1,0 @@
-'block_year': ['232'],
- 'pf_no': ['222'], 
- 'basic_pay_salary': ['2322'], 
- 'name': ['dsds'], 
- 'designation':['dsdsd'], 
- 'department_info': ['ds'], 
- 'leave_availability': ['True', 'True'], 
- 'leave_start_date': ['2024-02-22'], 
- 'leave_end_date': ['2024-02-22'], 
- 'date_of_leave_for_family': ['2024-02-22'], 
- 'nature_of_leave': ['dsds'], 
- 'purpose_of_leave': ['dsdsd'], 
- 'hometown_or_not': ['True'], 
- 'place_of_visit': [''], 
- 'address_during_leave': ['full street address'], 
- 'details_of_family_members_already_done': ['sds', 'dsd', 'dsd'], 
- 'info_1_1': ['1'], 'info_1_2': ['dsds'], 'info_1_3': ['12'], 
- 'info_2_1': ['2'], 'info_2_2': ['sds'], 'info_2_3': ['121'], 
- 'info_3_1': ['3'], 'info_3_2': ['dsds'], 'info_3_3': ['21'],
-  'info_4_1': [''], 'info_4_2': [''], 'info_4_3': [''], 
-  'info_5_1': [''], 'info_5_2': [''], 'info_5_3': [''], 
-  'info_6_1': [''], 'info_6_2': [''], 'info_6_3': [''], 
-  
-  'd_info_1_1': ['1'], 'd_info_1_2': ['sds'], 'd_info_1_3': ['21'], 'd_info_1_4': ['sdd'], 
-  
-  'd_info_2_1': ['2'], 'd_info_2_2': ['dsd'], 'd_info_2_3': ['23'], 'd_info_2_4': ['sds'],
-   'd_info_3_1': ['3'], 'd_info_3_2': ['sd'], 'd_info_3_3': ['21'], 'd_info_3_4': ['dds'], 
-   'd_info_4_1': [''], 'd_info_4_2': [''], 'd_info_4_3': [''], 'd_info_4_4': [''], 
-   'd_info_5_1': [''], 'd_info_5_2': [''], 'd_info_5_3': [''], 'd_info_5_4': [''], 
-   'd_info_6_1': [''], 'd_info_6_2': [''], 'd_info_6_3': [''], 'd_info_6_4': [''], 
-   'amount_of_advance_required': ['211'], 
-   'certified_family_dependents': ['dqwd'], 
-   'certified_advance': ['dqwd'], 
-   'adjusted_month': ['qwdwd'], 
-   'date': ['2024-02-22'], 
-   'phone_number_for_contact': ['2312123']
