+++ FusionIIIT/applications/hr2/tests/__init__.py
@@ -1,0 +1,1 @@
+"""Test package for HR2 module."""
